refactor(db): move relay invite store ownership
